render markdown in breadcrumbs
